+++ tests/provider/dwd/observation/test_api_data.py
@@ -692,6 +692,63 @@ def test_dwd_observation_data_result_tidy_metric():
     )
 
 
+@pytest.mark.remote
+def test_dwd_observation_data_10_minutes_result_tidy():
+    """ Test for actual values (tidy) in metric units """
+    request = DwdObservationRequest(
+        parameter=[
+            DwdObservationDatasetTree.MINUTE_10.TEMPERATURE_AIR.PRESSURE_AIR_SITE
+        ],
+        resolution=DwdObservationResolution.MINUTE_10,
+        start_date="1999-12-31 22:00",
+        end_date="1999-12-31 23:00",
+        tidy=True,
+        humanize=False,
+        si_units=False,
+    ).filter_by_station_id(
+        station_id=(1048,),
+    )
+
+    df = request.values.all().df
+
+    assert_frame_equal(
+        df,
+        pd.DataFrame(
+            {
+                "station_id": pd.Categorical(["01048"] * 7),
+                "dataset": pd.Categorical(["temperature_air"] * 7),
+                "parameter": pd.Categorical(["pp_10"] * 7),
+                "date": [
+                    datetime(1999, 12, 31, 22, 00, tzinfo=pytz.UTC),
+                    datetime(1999, 12, 31, 22, 10, tzinfo=pytz.UTC),
+                    datetime(1999, 12, 31, 22, 20, tzinfo=pytz.UTC),
+                    datetime(1999, 12, 31, 22, 30, tzinfo=pytz.UTC),
+                    datetime(1999, 12, 31, 22, 40, tzinfo=pytz.UTC),
+                    datetime(1999, 12, 31, 22, 50, tzinfo=pytz.UTC),
+                    datetime(1999, 12, 31, 23, 00, tzinfo=pytz.UTC),
+                ],
+                "value": pd.to_numeric(
+                    [
+                        996.1,
+                        996.2,
+                        996.2,
+                        996.2,
+                        996.3,
+                        996.4,
+                        pd.NA,
+                    ],
+                    errors="coerce",
+                ).astype(float),
+                "quality": pd.to_numeric(
+                    [1, 1, 1, 1, 1, 1, pd.NA], errors="coerce"
+                ).astype(float),
+            },
+        ),
+        # Needed since pandas 1.2?
+        check_categorical=False,
+    )
+
+
 def test_create_humanized_column_names_mapping():
     """ Test for function to create a mapping to humanized column names """
     kl_daily_hcnm = {