+++ wetterdienst/provider/dwd/observation/api.py
@@ -7,6 +7,7 @@ from itertools import repeat
 from typing import Dict, List, Optional, Union
 
 import pandas as pd
+from pandas import Timedelta, Timestamp
 
 from wetterdienst.core.scalar.request import ScalarRequestCore
 from wetterdienst.core.scalar.values import ScalarValuesCore
@@ -142,7 +143,6 @@ class DwdObservationValues(ScalarValuesCore):
                 and period == Period.HISTORICAL
             ):
                 date_ranges = self._get_historical_date_ranges(station_id, dataset)
-
                 for date_range in date_ranges:
                     periods_and_date_ranges.append((period, date_range))
             else:
@@ -205,7 +205,34 @@ class DwdObservationValues(ScalarValuesCore):
 
             parameter_df = parameter_df.append(period_df)
 
+        # TODO: temporary fix -> reduce time steps before 2000 by 1 hour
+        #  for 1minute and 10minutes resolution data
+        if not parameter_df.empty:
+            if self.stations.resolution in (Resolution.MINUTE_1, Resolution.MINUTE_10):
+                # Have to parse dates here although they should actually be parsed
+                # later on in the core API
+                parameter_df[Columns.DATE.value] = pd.to_datetime(
+                    parameter_df[Columns.DATE.value], format=self._datetime_format
+                )
+                parameter_df = self._fix_timestamps(parameter_df)
+
         return parameter_df
+
+    @staticmethod
+    def _fix_timestamps(df: pd.DataFrame) -> pd.DataFrame:
+        """
+        This method is used to fix timestamps for 10 minute resolution data provided
+        by DWD. The original document states
+        " The time stamp before the year 2000 is given in MEZ, the time stamp after
+        the year 2000 is given in UTC. "
+        :param df: pandas DataFrame with original timestamps
+        :return: pandas DataFrame with fixed timestamps
+        """
+        before_2000 = df[Columns.DATE.value].dt.year < 2000
+
+        df.loc[before_2000, Columns.DATE.value] -= Timedelta(hours=1)
+
+        return df
 
     def _tidy_up_df(self, df: pd.DataFrame, dataset) -> pd.DataFrame:
         """
@@ -348,6 +375,8 @@ class DwdObservationValues(ScalarValuesCore):
             dataset, self.stations.resolution, Period.HISTORICAL
         )
 
+        file_index = file_index[(file_index[Columns.STATION_ID.value] == station_id)]
+
         # The request interval may be None, if no start and end date
         # is given but rather the entire available data is queried.
         # In this case the interval should overlap with all files
@@ -402,7 +431,10 @@ class DwdObservationRequest(ScalarRequestCore):
         :return:
         """
         if self.start_date:
-            return pd.Interval(self.start_date, self.end_date, closed="both")
+            # cut of hours, seconds,...
+            start_date = Timestamp(self.start_date.date()).tz_localize(self.tz)
+            end_date = Timestamp(self.end_date.date()).tz_localize(self.tz)
+            return pd.Interval(start_date, end_date, closed="both")
 
         return None
 
@@ -416,11 +448,17 @@ class DwdObservationRequest(ScalarRequestCore):
         """
         historical_end = self._now_local.replace(month=1, day=1)
 
-        if self.start_date < historical_end:
-            historical_begin = self.start_date.tz_convert(historical_end.tz)
-        else:
-            historical_begin = historical_end + pd.tseries.offsets.DateOffset(years=-1)
+        # if self.start_date < historical_end:
+        #     historical_begin = self.start_date.tz_convert(historical_end.tz)
+        # else:
+        #     historical_begin = historical_end + pd.tseries.offsets.DateOffset(years=-1)
 
+        # a year that is way before any data is collected
+        historical_begin = pd.Timestamp(
+            datetime(year=1678, month=1, day=1)
+        ).tz_localize(historical_end.tz)
+
+        # TODO: Use date
         historical_interval = pd.Interval(
             left=historical_begin, right=historical_end, closed="both"
         )
@@ -438,6 +476,7 @@ class DwdObservationRequest(ScalarRequestCore):
         recent_end = self._now_local.replace(hour=0, minute=0, second=0)
         recent_begin = recent_end - pd.Timedelta(days=500)
 
+        # TODO: use date
         recent_interval = pd.Interval(
             left=recent_begin, right=recent_end, closed="both"
         )