+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Development
 ***********
 
 - [FIX] Add missing positional dataset argument for _create_empty_station_parameter_df
+- Fix timestamps of 1 minute / 10 minutes DWD data introducing a gap hour at the end of year 1999
 
 0.22.0 (01.10.2021)
 *******************